// frontend/src/app/api/model/task.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TaskPriority } from './taskPriority';
import { TaskStatus } from './taskStatus';


export interface Task { 
    id: number;
    title: string;
    description?: string | null;
    status: TaskStatus;
    priority: TaskPriority;
    dueDate?: string | null;
    createdAt: string;
    updatedAt: string;
    completedAt?: string | null;
    archived: boolean;
    overdue: boolean;
}
export namespace Task {
}