// frontend/src/app/api/model/taskStatus.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const TaskStatus = {
    Todo: 'TODO',
    InProgress: 'IN_PROGRESS',
    Done: 'DONE',
} as const;
export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];